Refactor: Move DB implementation to FSM
